+++ src/constants.rs
@@ -1,9 +1,17 @@
+/// Value for one weapon.
 pub const ONE_WEAPON: i32 = 1;
+/// Value for one scrap metal.
 pub const ONE_SCRAP: i32 = ONE_WEAPON * 2;
+/// Value for one reclaimed metal.
 pub const ONE_REC: i32 = ONE_SCRAP * 3;
+/// Value for one refined metal.
 pub const ONE_REF: i32 = ONE_REC * 3;
 
+/// Symbol for one key.
 pub const KEY_SYMBOL: &str = "key";
+/// Symbol for multiple keys.
 pub const KEYS_SYMBOL: &str = "keys";
+/// Symbol for metal.
 pub const METAL_SYMBOL: &str = "ref";
+/// Symbol for empty currencies.
 pub const EMPTY_SYMBOL: &str = "nothing";