+++ src/traits.rs
@@ -1,3 +1,4 @@
+use std::fmt::Debug;
 use serde::{Serialize, de::DeserializeOwned};
 
-pub trait SerializeCurrencies: Sized + std::fmt::Debug + Serialize + DeserializeOwned {}
+pub trait SerializeCurrencies: Sized + Debug + Serialize + DeserializeOwned {}